refactor: rename StyledLeva to StyledRoot for consistency
